requirements.txt
# === Servidor e API ===
fastapi>=0.109.0
uvicorn[standard]>=0.23.0

# === Deep Learning e Processamento de Imagem ===
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0

# === Utilidades ===
tqdm>=4.66.0
requests>=2.31.0

# === Logging e Diagnóstico ===
rich>=13.7.0